# packages/contracts/openapi/v0.1/core-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hear Our Language Core Backend API",
    "version": "0.1.0",
    "description": "The only product API. Runs on the Ryzen AI 9 laptop and owns orchestration, RAG, SQLite, teaching state, and fallback decisions. Draft pending Agent B walkthrough."
  },
  "servers": [{ "url": "http://127.0.0.1:8000", "description": "Ryzen AI 9 laptop" }],
...
  "paths": {
...
    "/api/lessons/analyze": {
      "post": {
        "tags": ["lessons"],
        "operationId": "analyzeLesson",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["image", "language"],
                "properties": {
                  "image": { "type": "string", "contentEncoding": "binary" },
                  "language": { "type": "string", "enum": ["nan-TW"] },
                  "use_fixture_on_failure": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Lesson" },
          "400": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" },
          "504": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/lessons/{lesson_id}": {
      "parameters": [{ "$ref": "#/components/parameters/LessonId" }],
      "get": {
        "tags": ["lessons"],
        "operationId": "getLesson",
        "description": "Teacher/parent review endpoint. Student routes must not request or store this full Lesson object.",
        "responses": {
          "200": { "$ref": "#/components/responses/Lesson" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["lessons"],
        "operationId": "reviewLesson",
        "description": "Teacher/parent review endpoint; not available to the student-mode adapter.",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "topic": { "type": "string" },
                  "source_text": { "type": "string" },
                  "accessible_activity": { "type": "string" },
                  "review_status": { "type": "string", "enum": ["pending", "approved", "rejected"] }
                },
                "minProperties": 1
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Lesson" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/utterances/normalize": {
      "post": {
        "tags": ["utterances"],
        "operationId": "normalizeUtterance",
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["schema_version", "text", "lang"],
                "properties": {
                  "schema_version": { "const": "0.1.0" },
                  "text": { "type": "string", "minLength": 1 },
                  "lang": { "type": "string", "enum": ["nan-TW", "zh-TW"] },
                  "tailo_citation": { "type": ["string", "null"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Utterance" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/sessions": {
      "post": {
        "tags": ["sessions"],
        "operationId": "createSession",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["schema_version", "lesson_id"],
                "properties": {
                  "schema_version": { "const": "0.1.0" },
                  "lesson_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Session" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/sessions/{session_id}": {
      "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
      "get": {
        "tags": ["sessions"],
        "operationId": "getSession",
        "responses": {
          "200": { "$ref": "#/components/responses/Session" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/sessions/{session_id}/snapshot": {
      "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
      "get": {
        "tags": ["sessions"],
        "operationId": "getSessionSnapshot",
        "description": "Returns the current student-safe snapshot after a reconnect or revision conflict.",
        "responses": {
          "200": { "$ref": "#/components/responses/Session" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/sessions/{session_id}/turns": {
      "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
      "post": {
        "tags": ["sessions"],
        "operationId": "submitTurn",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/SessionRevision" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TurnSubmission" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/TurnResult" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/sessions/{session_id}/actions": {
      "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
      "post": {
        "tags": ["sessions"],
        "operationId": "submitStudentAction",
        "description": "Dispatches a student-safe control action. Answer submission is intentionally excluded and must use /turns with a transcript.",
        "parameters": [
          { "$ref": "#/components/parameters/RequestId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/SessionRevision" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StudentActionRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/StudentAction" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/sessions/{session_id}/events": {
      "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
      "get": {
        "tags": ["sessions"],
        "operationId": "streamSessionEvents",
        "description": "Replayable SSE stream. Send Last-Event-ID or after to receive only events after the last applied event. Event ids are monotonic and payloads are safe for the student client.",
        "parameters": [
          { "$ref": "#/components/parameters/LastEventId" }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events. Each event id is a durable event_id and each data field is a JSON SessionEvent envelope. Empty backlogs produce a heartbeat comment.",
            "headers": { "X-Request-ID": { "$ref": "#/components/headers/RequestId" } },
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" },
                "examples": {
                  "sessionAction": {
                    "summary": "Durable event with reconnect id",
                    "value": "id: 2\\nevent: session.action\\ndata: {\\\"schema_version\\\":\\\"0.1.0\\\",\\\"event_id\\\":2,\\\"session_id\\\":\\\"session_demo_001\\\",\\\"event\\\":\\\"session.action\\\",\\\"request_id\\\":\\\"00000000-0000-4000-8000-000000000402\\\",\\\"revision\\\":1,\\\"payload\\\":{}}\\n\\n"
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/sessions/{session_id}/summary": {
      "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
      "get": {
        "tags": ["sessions"],
        "operationId": "getSessionSummary",
        "responses": {
          "200": { "$ref": "#/components/responses/ObserverSessionSummary" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "headers": {
      "RequestId": { "description": "Correlation UUID echoed or generated by the receiving service.", "schema": { "type": "string", "format": "uuid" } }
    },
    "parameters": {
      "RequestId": { "name": "X-Request-ID", "in": "header", "required": false, "schema": { "type": "string", "format": "uuid" } },
      "LessonId": { "name": "lesson_id", "in": "path", "required": true, "schema": { "type": "string" } },
      "SessionId": { "name": "session_id", "in": "path", "required": true, "schema": { "type": "string" } },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "Stable key for safe retry. Reusing a key replays the original result; using it for another operation returns a conflict. If omitted, X-Request-ID is used as the key.",
        "schema": { "type": "string", "minLength": 1, "maxLength": 128 }
      },
      "SessionRevision": {
        "name": "X-Session-Revision",
        "in": "header",
        "required": false,
        "description": "Optional optimistic-concurrency revision. A stale value returns SESSION_REVISION_CONFLICT with the current revision.",
        "schema": { "type": "integer", "minimum": 0 }
      },
      "LastEventId": {
        "name": "Last-Event-ID",
        "in": "header",
        "required": false,
        "description": "Last durable SSE event applied by the client.",
        "schema": { "type": "integer", "minimum": 0 }
      }
    },
    "schemas": {
      "Lesson": { "$ref": "../../schemas/v0.1/lesson.schema.json" },
      "Utterance": { "$ref": "../../schemas/v0.1/utterance.schema.json" },
...
      "StudentActionRequest": { "$ref": "../../schemas/v0.1/student-action.schema.json#/$defs/control_request" },
      "TurnSubmission": { "$ref": "../../schemas/v0.1/student-action.schema.json#/$defs/answer_submission" },
...
      "Error": { "$ref": "../../schemas/v0.1/error.schema.json" },
...
    },
    "responses": {
      "Lesson": {
        "description": "Canonical teacher-reviewable lesson. `answer_evidence` is teacher/parent-only and must never be returned by student routes.",
        "headers": { "X-Request-ID": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Lesson" } } }
      },
      "Utterance": {
        "description": "Normalized utterance.",
        "headers": { "X-Request-ID": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Utterance" } } }
      },
      "TurnResult": {
        "description": "Student-safe turn result.",
        "headers": { "X-Request-ID": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TurnResult" } } }
      },
      "StudentAction": {
        "description": "Student-safe session projection after a control action.",
        "headers": { "X-Request-ID": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StudentAction" } } }
      },
...
      "Session": {
        "description": "Student-safe session state; no answer keys or confidence fields.",
        "headers": { "X-Request-ID": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Session" } } }
      },
      "Error": {
        "description": "Stable error envelope.",
        "headers": { "X-Request-ID": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  }
}